+++ app/components/tack/poster/Poster.tsx
@@ -1,0 +1,20 @@
+import React from 'react';
+
+type PosterProps = {
+    
+
+
+
+    
+};
+
+const Poster:React.FC<PosterProps> = () => {
+    
+    return <div>
+
+
+
+
+    </div>
+}
+export default Poster;
